8.1 DB RelationShip

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from '../models/Video';
+import User from '../models/User';
 
 export const home = async (req, res) => {
   try {
@@ -15,16 +16,26 @@ export const getUpload = (req, res) => {
 
 export const postUpload = async (req, res) => {
   const {
+    session: {
+      user: { _id },
+    },
     body: { title, description, hashtags },
     file: { path: fileUrl },
   } = req;
+
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+
+    user.videos.push(newVideo._id);
+    user.save();
+
     return res.redirect('/');
   } catch (error) {
     return res.status(400).render('upload', {
@@ -36,46 +47,78 @@ export const postUpload = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate('owner');
+
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found' });
   }
+
   return res.render('watch', { pageTitle: video.title, video });
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
+
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found' });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   return res.render('edit', { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ _id: id });
+  const video = await Video.findById(id);
+
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found' });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   await Video.findByIdAndUpdate(id, {
     title,
     description,
     hashtags: formatHashtags(hashtags),
   });
+
   return res.redirect(`/videos/${id}`);
 };
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   await Video.findOneAndDelete(id);
+
   return res.redirect('/');
 };
 
 export const search = async (req, res) => {
   const { keyword } = req.query;
   let videos = [];
+
   if (keyword) {
     videos = await Video.find({
       title: {
@@ -83,5 +126,6 @@ export const search = async (req, res) => {
       },
     });
   }
+
   return res.render('search', { pageTitle: 'Search', videos });
 };
